Middle Raw MainView
I have made changes in the MainView. I added in the middle row 3 grid colomns and those will be showing data of the current laptop you are inspecting...

I made the edges of the grids smooth so the whole design gets a smooth look.

I added a SerClasses class and inside this class I will include every class that will be used more than once, like the Laptop class array I made in the class Data. It contains the checklists of the laptop.