Product code functioanlity
I finished the code I started earlier where I made an array of products and I wanted the laptops that are in the database to apear automaticly. I did this in here I made all the Laptop products in the Data class apear using the UIElementsManagement, this class will handel everything related to UI managment.